requirements.txt
psycopg2-binary==2.9.10
mcp==1.9.2

